# apps/web/src/lib/service-clients/service-properties/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "properties_service",
    "description": "",
    "termsOfService": "https://macro.com/terms",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/properties/definitions": {
      "get": {
        "tags": ["Properties"],
        "summary": "List property definitions with flexible filtering",
        "operationId": "list_properties",
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "description": "Filter by scope: 'user', 'team', 'system', or 'all'",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PropertyScope"
            }
          },
          {
            "name": "include_options",
            "in": "query",
            "description": "Whether to include property options in the response",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "for_entity_type",
            "in": "query",
            "description": "Filter properties applicable to a specific entity type",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/EntityType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Properties retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PropertyDefinitionResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": ["Properties"],
        "summary": "Create a new property definition",
        "operationId": "create_property_definition",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyDefinitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Property definition created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDefinition"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "403": {
            "description": "Team membership required for team scope"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/properties/definitions/{definition_id}": {
      "get": {
        "tags": ["Properties"],
        "summary": "Get a property definition by ID",
        "operationId": "get_property_definition",
        "parameters": [
          {
            "name": "definition_id",
            "in": "path",
            "description": "Property definition ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Property definition retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDefinition"
                }
              }
            }
          },
          "400": {
            "description": "Invalid property ID"
          },
          "404": {
            "description": "Property definition not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": ["Properties"],
        "summary": "Delete a property definition",
        "operationId": "delete_property_definition",
        "parameters": [
          {
            "name": "definition_id",
            "in": "path",
            "description": "Property definition ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Property definition deleted successfully"
          },
          "400": {
            "description": "Invalid property ID"
          },
          "403": {
            "description": "Cannot modify system properties"
          },
          "404": {
            "description": "Property definition not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePropertyDefinitionRequest": {
        "type": "object",
        "description": "Request to create a new property definition.",
        "required": ["scope", "display_name", "data_type"],
        "properties": {
          "data_type": {
            "$ref": "#/components/schemas/PropertyDataType"
          },
          "display_name": {
            "type": "string"
          },
          "scope": {
            "$ref": "#/components/schemas/CreatePropertyScope"
          }
        }
      },
      "CreatePropertyScope": {
        "type": "string",
        "description": "Ownership scope a client may request when creating a property definition.\nThe owning user or team is derived server-side from the authenticated caller -\nclients never supply owner ids. System properties are not creatable via the API.",
        "enum": ["user", "team"]
      },
      "DataType": {
        "type": "string",
        "description": "Data type for property values, determining storage and validation.",
        "enum": [
          "BOOLEAN",
          "DATE",
          "NUMBER",
          "STRING",
          "SELECT_NUMBER",
          "SELECT_STRING",
          "TAG",
          "ENTITY",
          "LINK"
        ]
      },
...
      "EntityType": {
        "type": "string",
        "description": "Type of entity that can be referenced by entity properties.",
        "enum": [
          "CALENDAR_EVENT",
          "CALL_RECORD",
          "CHANNEL",
          "CHAT",
          "COMPANY",
          "DOCUMENT",
          "PROJECT",
          "TASK",
          "THREAD",
          "USER"
        ]
      },
...
      "PropertyDataType": {
        "oneOf": [
          {
            "type": "object",
            "description": "Boolean true/false values.",
            "required": ["type"],
            "properties": {
              "type": {
                "type": "string",
                "enum": ["boolean"]
              }
            }
          },
          {
            "type": "object",
            "description": "Date and time values.",
            "required": ["type"],
            "properties": {
              "type": {
                "type": "string",
                "enum": ["date"]
              }
            }
          },
          {
            "type": "object",
            "description": "Numeric values.",
            "required": ["type"],
            "properties": {
              "type": {
                "type": "string",
                "enum": ["number"]
              }
            }
          },
          {
            "type": "object",
            "description": "String/text values.",
            "required": ["type"],
            "properties": {
              "type": {
                "type": "string",
                "enum": ["string"]
              }
            }
          },
          {
            "type": "object",
            "description": "Select property with numeric options.",
            "required": ["options", "multi", "type"],
            "properties": {
              "multi": {
                "type": "boolean"
              },
              "options": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/SelectNumberOption"
                }
              },
              "type": {
                "type": "string",
                "enum": ["select_number"]
              }
            }
          },
          {
            "type": "object",
            "description": "Select property with string options.",
            "required": ["options", "multi", "type"],
            "properties": {
              "multi": {
                "type": "boolean"
              },
              "options": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/SelectStringOption"
                }
              },
              "type": {
                "type": "string",
                "enum": ["select_string"]
              }
            }
          },
          {
            "type": "object",
            "description": "Entity reference property.",
            "required": ["multi", "type"],
            "properties": {
              "multi": {
                "type": "boolean"
              },
              "specific_type": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/EntityType"
                  }
                ]
              },
              "type": {
                "type": "string",
                "enum": ["entity"]
              }
            }
          },
          {
            "type": "object",
            "description": "Link property.",
            "required": ["multi", "type"],
            "properties": {
              "multi": {
                "type": "boolean"
              },
              "type": {
                "type": "string",
                "enum": ["link"]
              }
            }
          }
        ],
        "description": "Data type with embedded options for requests - provides compile-time validation."
      },
      "PropertyDefinition": {
        "type": "object",
        "description": "Property definition model (service representation).",
        "required": [
          "id",
          "owner",
          "display_name",
          "data_type",
          "is_multi_select",
          "created_at",
          "updated_at",
          "is_system",
          "is_metadata"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "data_type": {
            "$ref": "#/components/schemas/DataType"
          },
          "display_name": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_metadata": {
            "type": "boolean",
            "description": "Flag to indicate if this is a system-generated metadata property.\nNot stored in database - computed at service layer."
          },
          "is_multi_select": {
            "type": "boolean"
          },
          "is_system": {
            "type": "boolean",
            "description": "Flag to indicate if this is a system property (stored in DB)."
          },
          "owner": {
            "$ref": "#/components/schemas/PropertyOwner"
          },
          "specific_entity_type": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EntityType"
              }
            ]
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PropertyDefinitionResponse": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/PropertyDefinition"
          },
          {
            "$ref": "#/components/schemas/PropertyDefinitionWithOptions"
          }
        ],
        "description": "Response for property definition with optional property options"
      },
      "PropertyDefinitionWithOptions": {
        "type": "object",
        "description": "Property definition with its associated options (service representation).",
        "required": ["definition", "property_options"],
        "properties": {
          "definition": {
            "$ref": "#/components/schemas/PropertyDefinition"
          },
          "property_options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyOption"
            }
          }
        }
      },
...
      "PropertyOption": {
        "type": "object",
        "description": "A selectable option for select-type properties (service representation).",
        "required": [
          "id",
          "property_definition_id",
          "display_order",
          "value",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "color": {
            "type": ["string", "null"]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "display_order": {
            "type": "integer",
            "format": "int32"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "property_definition_id": {
            "type": "string",
            "format": "uuid"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "value": {
            "$ref": "#/components/schemas/PropertyOptionValue"
          }
        }
      },
...
      "PropertyOptionValue": {
        "oneOf": [
          {
            "type": "object",
            "description": "String value for SelectString properties",
            "required": ["value", "type"],
            "properties": {
              "type": {
                "type": "string",
                "enum": ["string"]
              },
              "value": {
                "type": "string",
                "description": "String value for SelectString properties"
              }
            }
          },
          {
            "type": "object",
            "description": "Number value for SelectNumber properties",
            "required": ["value", "type"],
            "properties": {
              "type": {
                "type": "string",
                "enum": ["number"]
              },
              "value": {
                "type": "number",
                "format": "double",
                "description": "Number value for SelectNumber properties"
              }
            }
          }
        ],
        "description": "The value of a property option - either a string or a number."
      },
      "PropertyOwner": {
        "oneOf": [
          {
            "type": "object",
            "description": "User-scoped property.",
            "required": ["user_id", "scope"],
            "properties": {
              "scope": {
                "type": "string",
                "enum": ["user"]
              },
              "user_id": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "description": "Team-scoped property.",
            "required": ["team_id", "scope"],
            "properties": {
              "scope": {
                "type": "string",
                "enum": ["team"]
              },
              "team_id": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "type": "object",
            "description": "System-owned property (no user or team owner).",
            "required": ["scope"],
            "properties": {
              "scope": {
                "type": "string",
                "enum": ["system"]
              }
            }
          }
        ],
        "description": "Defines who owns a property - user-scoped, team-scoped, or system."
      },
      "PropertyScope": {
        "type": "string",
        "description": "Scope filter for property queries",
        "enum": ["user", "team", "system", "all"]
      },
...
      "SelectNumberOption": {
        "type": "object",
        "description": "Option for select number properties in requests.",
        "required": ["display_order", "value"],
        "properties": {
          "display_order": {
            "type": "integer",
            "format": "int32"
          },
          "value": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "SelectStringOption": {
        "type": "object",
        "description": "Option for select-type properties in requests.",
        "required": ["display_order", "value"],
        "properties": {
          "display_order": {
            "type": "integer",
            "format": "int32"
          },
          "value": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}